ETL-206 Use Poetry to install Sentry as dependency
