+++ .gitignore
@@ -30,3 +30,6 @@ override.tf.json
 
 # JetBrains IDEs
 .idea
+
+# Build directory
+/build/
