+++ lambda.tf
@@ -6,14 +6,20 @@ resource "aws_lambda_permission" "apigw_lambda" {
   source_arn    = "arn:aws:execute-api:${var.region}:${data.aws_caller_identity.current.account_id}:${aws_api_gateway_rest_api.sftp.id}/*/${aws_api_gateway_method.get.http_method}${aws_api_gateway_resource.config.path}"
 }
 
+locals {
+  deployment_archive_path = "${path.module}/build.zip"
+}
+
+
 resource "aws_lambda_function" "sftp" {
+  filename         = local.deployment_archive_path
+  source_code_hash = filebase64sha256(local.deployment_archive_path)
+
   description      = "A function to lookup and return user data from AWS Secrets Manager."
-  filename         = data.archive_file.sftp_lambda.output_path
   function_name    = "${var.name_prefix}-sftp-transfer-server-custom-idp-lambda${var.name_suffix}"
   role             = aws_iam_role.sftp_lambda_role.arn
-  handler          = "sftp_lambda.lambda_handler"
-  runtime          = "python3.7"
-  source_code_hash = filebase64sha256(data.archive_file.sftp_lambda.output_path)
+  handler          = "transfer_server_custom_idp.lambda_handler.lambda_handler"
+  runtime          = "python3.8"
   reserved_concurrent_executions = 5
 
   tracing_config {
@@ -86,11 +92,4 @@ EOF
 resource "aws_iam_role_policy_attachment" "sftp_lambda_role" {
   role       = aws_iam_role.sftp_lambda_role.name
   policy_arn = "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole"
-}
-
-
-data "archive_file" "sftp_lambda" {
-  type        = "zip"
-  source_file = "${path.module}/files/sftp_lambda.py"
-  output_path = "${path.module}/files/lambda.zip"
 }
