+++ pyproject.toml
@@ -1,0 +1,29 @@
+[tool.poetry]
+name = "transfer-server-custom-idp"
+version = "0.0.2"
+description = "Secrets Manager identity provider for AWS Transfer Family"
+authors = [
+    "Anatoly Scherbakov <altaisoft@gmail.com>"
+]
+
+[tool.poetry.dependencies]
+python = ">=3.8,<3.9"
+sentry-sdk = "^1.0.0"
+Jinja2 = "^3.0.1"
+pydantic = "^1.8.2"
+
+
+[tool.poetry.dev-dependencies]
+wemake-python-styleguide = "^0.15.2"
+flakehell = "^0.9.0"
+pip = "^21.0.1"
+mypy = "^0.812"
+flake8 = "^3.9.0,<3.9.1"
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
+
+
+[tool.flakehell]
+base = "https://raw.githubusercontent.com/Recall-Masters/standards/main/pyproject.toml"
